add teacherName to news

+++ src/app/services/news-service/news.service.ts
@@ -17,7 +17,7 @@ export class NewsService {
 
   getNewsByTeacherReference(teacher: DocumentReference) {
     return this.firestore.collection(teacher.parent.path).doc(teacher.id)
-      .collection('News', query => query.orderBy('start_time').orderBy('topic')).snapshotChanges()
+      .collection('News', query => query.orderBy('start_time', 'desc').orderBy('topic')).snapshotChanges()
       .pipe(map(result => result.map(item => item.payload.doc).map(doc => {
         return { id: doc.id, ref: doc.ref, ...doc.data() as News } as News;
       })));
@@ -37,15 +37,15 @@ export class NewsService {
     return this.firestore.collection(newsRef.parent.path).doc(newsRef.id).update(Object.assign({}, news));
   }
 
-  getNews(news_id: string) {
-    return this.firestore.collection('News').doc(news_id).snapshotChanges();
-  }
-
-  deleteNews(news_id: string) {
-    this.getNews(news_id).subscribe(result => {
-      let news = result.payload.data() as News;
-      this.afStorage.ref(news.image).delete();
-    });
-    this.firestore.collection('News').doc(news_id).delete();
+  deleteNews(news: News) {
+    try {
+      if (news.image) {
+        this.afStorage.ref(news.image).delete();
+      }
+      this.firestore.doc(news.ref.path).delete();
+    } catch (error) {
+      console.error(error);
+      throw new Error('เกิดข้อมผิดพลาดในการลบ กรุณาลองใหม่อีกครั้งภายหลัง');
+    }
   }
 }
